refactor(web): split earn motion components
